Structure the repo and start of backend
This commit structures the repo in backend, frontend, and feedback. The
respective folders hold the current work done. Backend is only a minimal
prototype that is not yet tested nor working. It just serves as a
example to start the journey to implement a real backend into the urban
gardening project. urban-gardening was renamed to frontend to better
show what it is used for. Same holds true for the feedback folder which
now contains the feedback loop together with creating the mock data to
be written into the influxdb.